Trait libnoentiendo::cpu::MemoryIO

pub trait MemoryIO {
    fn read(&mut self, address: u16) -> u8;
    fn write(&mut self, address: u16, value: u8);
    fn read_word(&mut self, address: u16) -> u16;
    fn write_word(&mut self, address: u16, value: u16);
}

Read and write from the system’s memory.

Required Methods

fn read(&mut self, address: u16) -> u8

Read a byte from the given address in memory.

fn write(&mut self, address: u16, value: u8)

Write a byte to the given address in memory.

fn read_word(&mut self, address: u16) -> u16

Read a word (little-endian) from the given address in memory.

fn write_word(&mut self, address: u16, value: u16)

Write a word (little-endian) to the given address in memory.