Module libnoentiendo::keyboard

Structs

KeyState

A set of keys that are currently pressed. Parameter T is the type of the key symbols.

SymbolAdapter

An adapter that maps physical key positions to symbols.

Enums

KeyMappingStrategy

Represents different approaches to mapping key states, to allow the user to indicate their preference.

KeyPosition

A representation for a position on a modern keyboard. Source: https://en.wikipedia.org/wiki/Keyboard_layout#/media/File:Qwerty.svg

KeySymbol

A representation for a symbol on a modern keyboard.

Traits

KeyAdapter

Represents a mapping from a key state of one type to a key state of another type. Mappings can be symbolic (preserve symbols across the mapping, and rewrite modifier keys as needed) or physical (maintain a one-to-one mapping from physical keys to physical keys).