Module libnoentiendo::keyboard::commodore

Keys and adapters for the Commodore 64, VIC-20, and other Commodore machines.

Structs

C64KeyboardAdapter

An adapter for mapping positions on a standard keyboard to keys on a VIC-20 or Commodore 64.

C64SymbolAdapter

An adapter for mapping symbols pressed on a standard keyboard to keys pressed on a Commodore 64.

Enums

C64Keys

Keys found on a VIC-20 or Commodore 64 keyboard.