Module libnoentiendo::memory

A memory::Memory implementation can be read from and written to, but it can also be polled for interrupts. This is used for the PIA, VIA, and other chips that interface over memory but also trigger interrupts. The memory module provides implementations for various types of memory and other memory-mapped devices. Mappings are handled using memory::BranchMemory.

Off-cycle memory access schemes such as the one used in the VIC-20 are handled using the memory::DMA trait. Systems may have objects which implement this trait attached, and these objects will have a chance to access memory in between clock cycles.

Re-exports

pub use ports::NullPort;

pub use ports::Port;

Modules

mos652x

ports

Structs

BankedMemory

BlockMemory

Represents a simple block of contiguous memory, with no additional hardware. This can be used to represent both RAM and ROM. Reading from this memory is side-effect free.

BranchMemory

Maps several Memory objects into a single contiguous address space. Each mapped object is assigned a starting address, and reads and writes will have the starting address subtracted from them before being passed to the underlying Memory object.

Mos6510Port

Represents the port built into a MOS 6510 processor, mapped to memory addresses 0x0000 (for the DDR) and 0x0001 (for the port itself).

NullMemory

Memory that does nothing when read or written to.

SystemInfo

Information about the system that Memory implementations can use to determine if an interrupt should be triggered.

Enums

ActiveInterrupt

Represents the state of the interrupts on the system.

Traits

DMA

Represents a system component which may access the memory in between cycles. This is used to implement the VIC chip’s alternate-clock memory access.

Memory

Represents a contiguous block of memory which can be read, written, reset, and polled to see if an interrupt has been triggered.