Trait libnoentiendo::memory::ports::Port

pub trait Port {
    fn read(&mut self) -> u8;
    fn write(&mut self, value: u8);
    fn poll(&mut self, info: &SystemInfo) -> bool;
    fn reset(&mut self);
}

A Port that can be read from, written to, reset, or polled for interrupts. Used in the MOS 6520 PIA and the 6522 VIA.

Required Methods

fn read(&mut self) -> u8

Read a byte from the port. This is implementation-defined, and may have side effects.

fn write(&mut self, value: u8)

Write a byte to the port. This is implementation-defined.

fn poll(&mut self, info: &SystemInfo) -> bool

Poll the port for interrupts. A port may trigger an interrupt for any implementation-defined reason.

fn reset(&mut self)

Reset the port to its initial state, analogous to a system reboot.