Trait libnoentiendo::memory::DMA

pub trait DMA {
    fn dma(&mut self, memory: &mut Box<dyn Memory>, info: &SystemInfo);
}

Represents a system component which may access the memory in between cycles. This is used to implement the VIC chip’s alternate-clock memory access.

Required Methods

fn dma(&mut self, memory: &mut Box<dyn Memory>, info: &SystemInfo)

Handle DMA for the current instruction tick. Implementations may perform whatever operations on the memory that they would like.