Trait libnoentiendo::memory::Memory

pub trait Memory {
    fn read(&mut self, address: u16) -> u8;
    fn write(&mut self, address: u16, value: u8);
    fn reset(&mut self);
    fn poll(&mut self, info: &SystemInfo) -> ActiveInterrupt;
}

Represents a contiguous block of memory which can be read, written, reset, and polled to see if an interrupt has been triggered.

Required Methods

fn read(&mut self, address: u16) -> u8

Read a byte from this memory at the given address. Implementations may trigger side effects as a result of this read.

fn write(&mut self, address: u16, value: u8)

Write a byte to this memory at the given address.

fn reset(&mut self)

Reset this memory to its initial state, e.g. after a system reboot. Sometimes this will clear the contents of the memory, like with RAM. Other times this is a no-op, e.g. for ROM.

fn poll(&mut self, info: &SystemInfo) -> ActiveInterrupt

Poll this memory to see if an interrupt has been triggered. Implementations may trigger an NMI or IRQ for any implementation-dependent reason.