Module libnoentiendo::platform

A platform::Platform consumes a system and runs it. Platforms provide access to the video output, keyboard input, system random number generator, and other details via a platform::PlatformProvider. Some platforms run synchronously (taking over the thread) while others run asynchronously with the help of an event loop (such as when compiling to WASM). Platforms are defined in the platform module. Currently, available platforms include TextPlatform for simple headless text-based operation, WinitPlatform for a graphical window on a desktop environment, and CanvasPlatform for drawing to a <canvas> element on the web. In the future, platforms for mobile apps are planned, in addition to a platform for running on a microcontroller (e.g. the RP2040).

Structs

Color

Represents an RGB color with 8 bits per channel.

JoystickState

Represents the current state of the connected joystick.

TextPlatform

Represents a platform which exclusively operates over text mode, without any visible graphical output. This reads from and writes to the terminal. This platform runs synchronously.

TextPlatformProvider

WindowConfig

Represents the configuration of a GUI window that the system can request from the platform.

WinitPlatform

A platform implementation for desktop platforms using Winit and Pixels. This platform runs synchronously.

WinitPlatformProvider

Traits

AsyncPlatform

A platform which can be run asynchronously.

Platform

A Platform provides platform-specific functionality to the emulator. It handles starting and ticking the system, and provides a PlatformProvider to the system for screen/keyboard/etc. access.

PlatformProvider

SyncPlatform

A platform which can be run synchronously.