Module libnoentiendo::roms

ROM file loading and unloading is different on different platforms: desktop platforms typically load ROMs from a file, while WebAssembly platforms need to load ROMs from a Uint8Array. ROM file definition and loading is handled in the roms module, with specific roms::DiskLoadable and roms::JsValueLoadable traits for these two cases. Loaded ROMs are represented with a roms::RomFile object, which can be passed to memory::BlockMemory::from_file.

Structs

RomFile

Represents a predefined, immutable ROM file. Useful for storing character, BASIC, kernal, etc. ROMs.

Traits

DiskLoadable

Represents a value which can be loaded from a file.