Module libnoentiendo::systems

Systems are created by a systems::SystemFactory. A system is created with some roms, configuration, and platform. For instance, the create implementation on systems::PetSystemFactory takes in systems::pet::PetSystemRoms, systems::pet::PetSystemConfig, and an Arc<dyn PlatformProvider>.

Re-exports

pub use brooke::BrookeSystemFactory;

pub use c64::C64SystemConfig;

pub use c64::C64SystemFactory;

pub use easy::EasySystemFactory;

pub use klaus::KlausSystemFactory;

pub use pet::PetSystemConfig;

pub use pet::PetSystemFactory;

pub use vic::Vic20SystemConfig;

pub use vic::Vic20SystemFactory;

pub use vic::Vic20SystemRoms;

Modules

brooke

c64

easy

klaus

pet

vic

Structs

C64SystemRoms

The set of ROM files required to run a Commodore 64 system.

PetSystemRoms

The set of ROM files required to run a PET system.

Traits

SystemFactory

A factory for creating a system from a set of ROM files and a platform.