Module libnoentiendo::systems::pet

Structs

PetPia1PortA

Port A on the first PIA. This is used for generating the 60Hz interrupt (which is fired when the screen drawing reaches the last line), and for setting the active row of the keyboard matrix.

PetPia1PortB

Port B on the first PIA. This is used for reading the keyboard matrix.

PetSystem

The Commodore PET system.

PetSystemConfig

Configuration for a Commodore PET system.

PetSystemRoms

The set of ROM files required to run a PET system.

Enums

PetKeys

The keys found on the PET’s “Graphics” keyboard. Source: https://commons.wikimedia.org/wiki/File:PET_Keyboard.svg