Struct libnoentiendo::systems::pet::PetPia1PortA

pub struct PetPia1PortA { /* private fields */ }

Port A on the first PIA. This is used for generating the 60Hz interrupt (which is fired when the screen drawing reaches the last line), and for setting the active row of the keyboard matrix.

Implementations

impl PetPia1PortA

pub fn new() -> Self

pub fn get_keyboard_row(&self) -> Rc<Cell<u8>>