Module libnoentiendo::keyboard

Various representations of keyboard scancodes are required in different parts of the codebase. Each platform typically has its own definition of a scancode (e.g. JavaScript’s event.code or Winit’s VirtualKeyCode), and each emulated system has a different set of keys (e.g. the Commodore key on the VIC-20 or the standalone " key on the PET).

Utilities for handling keyboard input are defined in the keyboard module. This module defines a keyboard::KeyPosition enum to represent the physical keys which appear on a modern host keyboard and a keyboard::KeySymbol enum to represent the symbols which can be typed with a modern host keyboard. It also defines a keyboard::KeyState<T> struct to represent the set of currently-pressed keys in some representation T.

Mapping between different keyboard representations is handled using keyboard::KeyAdapter<F, T> implementations. These exist for four different scenarios:

Mapping from platform-specific scancodes to the common keyboard::KeyPosition representation

Mapping from keyboard::KeyPosition to system-specific scancodes, using a one-to-one “physical” mapping strategy

Mapping from keyboard::KeyPosition to keyboard::KeySymbol, respecting the currently-pressed modifier keys

Mapping from keyboard::KeySymbol to system-specific scancodes, to preserve the symbols that the user pressed even if it requires rewriting the currently-pressed modifier keys

Modules

commodore

Keys and adapters for the Commodore 64, VIC-20, and other Commodore machines.

Structs

KeyState

A set of keys that are currently pressed. Parameter T is the type of the key symbols.

SymbolAdapter

An adapter that maps physical key positions to symbols.

Enums

KeyMappingStrategy

Represents different approaches to mapping key states, to allow the user to indicate their preference.

KeyPosition

A representation for a position on a modern keyboard. Source: https://en.wikipedia.org/wiki/Keyboard_layout#/media/File:Qwerty.svg

KeySymbol

A representation for a symbol on a modern keyboard.

VirtualKey

Traits

KeyAdapter

Represents a mapping from a key state of one type to a key state of another type. Mappings can be symbolic (preserve symbols across the mapping, and rewrite modifier keys as needed) or physical (maintain a one-to-one mapping from physical keys to physical keys).