Trait libnoentiendo::keyboard::KeyAdapter

pub trait KeyAdapter<F: PartialEq + Clone, T: PartialEq + Clone> {
    // Required method
    fn map(state: &KeyState<F>) -> KeyState<T>;
}

Represents a mapping from a key state of one type to a key state of another type. Mappings can be symbolic (preserve symbols across the mapping, and rewrite modifier keys as needed) or physical (maintain a one-to-one mapping from physical keys to physical keys).

Required Methods

fn map(state: &KeyState<F>) -> KeyState<T>

Map the current state of the keyboard with symbols of type F to an equivalent keyboard state with symbols of type T.

Object Safety

This trait is not object safe.