Module libnoentiendo::memory

Re-exports

pub use ports::NullPort;

pub use ports::Port;

Modules

pia

ports

via

Structs

BlockMemory

Represents a simple block of contiguous memory, with no additional hardware. This can be used to represent both RAM and ROM. Reading from this memory is side-effect free.

BranchMemory

Maps several Memory objects into a single contiguous address space. Each mapped object is assigned a starting address, and reads and writes will have the starting address subtracted from them before being passed to the underlying Memory object.

NullMemory

Memory that does nothing when read or written to.

SystemInfo

Information about the system that Memory implementations can use to determine if an interrupt should be triggered.

Enums

ActiveInterrupt

Represents the state of the interrupts on the system.

Traits

DMA

Represents a system component which may access the memory in between cycles. This is used to implement the VIC chip’s alternate-clock memory access.

Memory

Represents a contiguous block of memory which can be read, written, reset, and polled to see if an interrupt has been triggered.