Module libnoentiendo::memory::mos652x

The various interface adapters (6520, 6522, 6526) for the MOS 6502 CPU.

Structs

Cia

The MOS 6526 Complex Interface Adapter (CIA). Contains two ports, two timers, a real-time clock, a shift register, and interrupt registers.

InterruptRegister

Registers for interrupt flags and interrupt enable bits. Each bit from 0 to 6 corresponds to an interrupt source.

Pia

The MOS 6520 Peripheral Interface Adapter (PIA), containing two ports and some control lines.

PortRegisters

A port and its associated registers on the MOS 6522 VIA or MOS 6526 CIA.

ShiftRegister

The shift register used by the MOS 6522 VIA and MOS 6526 CIA.

Timer

A timer circuit on the MOS 6522 VIA or MOS 6526 CIA.

Via

The MOS 6522 Versatile Interface Adapter (VIA). Contains two ports, two timers, a shift register, and some interrupt and control registers. Source: http://archive.6502.org/datasheets/mos_6522_preliminary_nov_1977.pdf

Enums

TimerClockSource

The source of the timer’s clock, which controls the rate at which its clock decrements.

TimerOutput

The manner in which the timer will output signals to the port, if at all.